fio: Fix hack around GCS storage issue
We have some happen on very rare occasions. Maybe twice a year or
about every 50k container pushes, we'll wind up with a GCS object like:

  gs://subscriber_registry/docker/registry/v2/repositories/andy-corp/shellhttpd/_manifests/revisions/sha256/d0f3dd07d5dfd10bf50d41892c3e23eac3c016f2652f91a597924d0481b4acf5/link

Being an empty file. I don't know if its a GCS bug or a registry bug.
However, it's not really a huge problem as the remainder of the
container image content is in-tact. By simply ignoring this condition we
can avoid a very tricky problem to diagnose.

This was tested/verified by setting the contents of the storage object
above to an empty file and then doing a local docker pull.

Signed-off-by: Andy Doan <andy@foundries.io>